Test turtle parsing
The current web-id spec recommends turtle and HTML RDFa parsing,
since I thought rdflib supported the formats, I suspected enabling
it should be pretty easy.

I needed to run a conversion tool to convert the rdf+xml files to turtle
and add some entries on the content-type maps for it to work.